Refactor: extract QuestionnaireSimulationController
Extracts the simulation methods out of QuestionnaireController in
order to have a distinct controller to run a questionnaire
simulation, focusing the original QuestionnaireController to CRUD
operations over questionnaires.
