+++ web/router.ex
@@ -114,10 +114,10 @@ defmodule Ask.Router do
         resources "/questionnaires", QuestionnaireController, except: [:new, :edit] do
           get "/export_zip", QuestionnaireController, :export_zip, as: :questionnaires_export_zip
           post "/import_zip", QuestionnaireController, :import_zip, as: :questionnaires_import_zip
-          post "/simulation", QuestionnaireController, :start_simulation, as: :questionnaires_start_simulation
-          post "/simulation/message", QuestionnaireController, :sync_simulation, as: :questionnaires_sync_simulation
-          get "/simulation/:respondent_id", QuestionnaireController, :get_last_simulation_response, as: :get_last_simulation_response
           put "/update_archived_status", QuestionnaireController, :update_archived_status, as: :update_archived_status
+          post "/simulation", QuestionnaireSimulationController, :start, as: :questionnaires_start_simulation
+          post "/simulation/message", QuestionnaireSimulationController, :sync, as: :questionnaires_sync_simulation
+          get "/simulation/:respondent_id", QuestionnaireSimulationController, :get_last_response, as: :get_last_simulation_response
         end
         get "/autocomplete_vars", ProjectController, :autocomplete_vars, as: :autocomplete_vars
         get "/autocomplete_primary_language", ProjectController, :autocomplete_primary_language, as: :autocomplete_primary_language
